IndividualDisplacements.jl

IndividualDisplacements.jl

IndividualDisplacements.jl computes individual displacments of elementary points over a gridded model domain. It can also read / write them from / to file. A typical application is the simulation and analysis of materials drifting or moving over the Global Ocean (e.g. plastics or planktons) or Atmosphere (e.g. dust or chemicals). Inter-operability with popular climate model grids and MeshArrays.jl is an important prospect. IndividualDisplacements.jl was initially designed in relation to MITgcm, ECCOv4 (Forget et al. 2015), and CBIOMES model simulations.

API Guide

IndividualDisplacements.PlotBasic

IndividualDisplacements.PlotMapProj

IndividualDisplacements.ReadDisplacements

IndividualDisplacements.ReadGriddedFields

IndividualDisplacements.VelComp

IndividualDisplacements.VelCopy

IndividualDisplacements.myread

IndividualDisplacements.PlotBasic — Method.

PlotBasic(df::DataFrame,nn::Integer)

Plot random subset of size nn trajectories.

source

IndividualDisplacements.PlotMapProj — Method.

PlotMapProj(df::DataFrame,nn::Integer)

Plot random subset of size nn trajectories using PyPlot & basemap.

source

IndividualDisplacements.ReadDisplacements — Method.

ReadDisplacements(dirIn::String,prec::DataType)

Read displacements from MITgcm output file using MeshArrays and return as a DataFrame.

source

IndividualDisplacements.ReadGriddedFields — Method.

ReadGriddedFields()

Read gridded variables from file using MeshArrays and return result in uvetc Dictionary.

source

IndividualDisplacements.VelComp — Method.

VelComp(du,u,p::Dict,tim)

Interpolate velocity from gridded fields and return position increment du

source

IndividualDisplacements.VelCopy — Method.

VelCopy(du,u,p::DataFrame,t)

Interpolate velocity from MITgcm float_trajectories output and return position increment du.

source

IndividualDisplacements.myread — Method.

myread()

Read a gridded variable from 2x2 tile files. This is used in ReadGriddedFields() with flt_example/

source