IndividualDisplacements.jl

IndividualDisplacements.jl computes elementary point displacements over a gridded Earth domain (e.g. a climate model C-grid). A typical application is the simulation and analysis of materials moving through atmospheric flows (e.g. dust or chemicals) or oceanic flows (e.g. plastics or planktons).

Inter-operability with popular climate model grids and their representation in MeshArrays.jl is a central element. The package can also read and plot trajectory simulation output from e.g. the MITgcm. It was originally developed using ECCOv4 and CBIOMES ocean model simulations (Forget et al. 2015).

The VelComp! and VelComp functions compute the velocity of tracked points. tests/runtests.jl uses solid body rotation as a benchmark (see below).

<img src="https://github.com/JuliaClimate/IndividualDisplacements.jl/blob/master/examples/SolidBodyRotation.png" width="50%">

API Guide

IndividualDisplacements.ReadDisplacements

IndividualDisplacements.VelComp

IndividualDisplacements.VelComp!

IndividualDisplacements.VelCopy

IndividualDisplacements.example1

IndividualDisplacements.example2

IndividualDisplacements.example2_setup

IndividualDisplacements.myread

IndividualDisplacements.ReadDisplacements — Method

ReadDisplacements(dirIn::String,prec::DataType)

Read displacements from MITgcm output file using MeshArrays and return as a DataFrame.

IndividualDisplacements.VelComp! — Method

VelComp!(du,u,p::Dict,tim)

Interpolate velocity from gridded fields (after exchange on u0,v0) and return position increment du (i.e. x,y,fIndex).

IndividualDisplacements.VelComp — Method

VelComp(du,u,p::Dict,tim)

Interpolate velocity from gridded fields and return position increment du

IndividualDisplacements.VelCopy — Method

VelCopy(du,u,p::DataFrame,t)

Interpolate velocity from MITgcm float_trajectories output and return position increment du.

IndividualDisplacements.example1 — Method

example1()

Global ocean case – just reading from file for now.

df=IndividualDisplacements.example1()

p=dirname(pathof(IndividualDisplacements))
include(joinpath(p,"plot_pyplot.jl"))
PyPlot.figure(); PlotMapProj(df,300); gcf()

IndividualDisplacements.example2 — Method

example2()

Reproducing MITgcm/verification/flt_example/ case. This is based on an extended and modified configuration of the standard MITgcm test case.

(df,ref,sol)=IndividualDisplacements.example2();

p=dirname(pathof(IndividualDisplacements))
include(joinpath(p,"plot_pyplot.jl"))
PyPlot.figure(); PlotBasic(df,300,100000.0); gcf()

using Plots
Plots.plot(sol[1,:],sol[2,:],linewidth=5,lc=:black, title="One Trajectory Example",
xaxis="x",yaxis="y",label="Julia Solution") # legend=false
pl=Plots.plot!(ref[1,:],ref[2,:],lw=3,ls=:dash,lc=:red,label="MITgcm Solution")

IndividualDisplacements.example2_setup — Method

example2_setup()

Read gridded variables from file using MeshArrays and return result in uvetc Dictionary.

IndividualDisplacements.myread — Method

myread()

Read a gridded variable from 2x2 tile files. This is used in example2_setup() with flt_example/