IndividualDisplacements.jl

IndividualDisplacements.jl computes elementary point displacements over a gridded Earth domain (e.g. a climate model C-grid). A typical application is the simulation and analysis of materials moving through atmospheric flows (e.g. dust or chemicals) or oceanic flows (e.g. plastics or planktons).

Inter-operability with common climate model grids and their representation in MeshArrays.jl is a central element. The package can also read and plot trajectory simulation output from e.g. the MITgcm. It was originally developed using ECCOv4 and CBIOMES ocean model simulations (Forget et al. 2015).

The ⬡ and ⬡! functions compute the tracked point / individual / agent velocities.

<img src="https://github.com/gaelforget/MarineEcosystemNotebooks/raw/master/figs/SolidBodyRotation.gif" width="40%"> <img src="https://github.com/gaelforget/MarineEcosystemNotebooks/raw/master/figs/RandomFlow.gif" width="40%">

tests/runtests.jl uses this solid body rotation case as a unit test case.

List Of Examples

A solid-body-rotation example is used for unit testing:

test/runtests.jl
examples/SolidBodyRotation.jl

Two examples using ⬡! and update_locations.jl:

examples/RandomFlow_fleet.jl
examples/GlobalDomain_fleet.jl

<img src="https://github.com/JuliaClimate/GlobalOceanNotebooks/raw/master/OceanTransports/LatLonCap300mDepth.png" width="80%">

Plotting recipes using three popular plotting packages:

examples/recipes_plots.jl
examples/recipes_makie.jl
examples/recipes_pyplot.jl

Three other examples using ⬡ are documented in the API Guide:

examples/examples123.jl
examples/example2fleet.jl
examples/example2more.jl

API Guide

IndividualDisplacements.NeighborTileIndices_cs

IndividualDisplacements.NeighborTileIndices_dpdo

IndividualDisplacements.RelocationFunctions_cs

IndividualDisplacements.RelocationFunctions_cs_check

IndividualDisplacements.VelComp

IndividualDisplacements.VelComp!

IndividualDisplacements.VelCopy

IndividualDisplacements.initialize_grid_locations

IndividualDisplacements.initialize_random_locations

IndividualDisplacements.postprocess_ODESolution

IndividualDisplacements.postprocess_ODESolution_simple

IndividualDisplacements.randn_lonlat

IndividualDisplacements.read_drifters

IndividualDisplacements.read_flt

IndividualDisplacements.read_mds

IndividualDisplacements.read_uvetc

IndividualDisplacements.read_uvetc

IndividualDisplacements.read_velocities

IndividualDisplacements.update_location_cs!

IndividualDisplacements.update_location_dpdo!

IndividualDisplacements.initialize_grid_locations — Function

initialize_grid_locations(uvetc::Dict,n_subset::Int=1)

Define initial condition (u0,du) as a subset of grid points

IndividualDisplacements.initialize_random_locations — Function

initialize_random_locations(Γ::Dict,n::Int=1 ; s=missing,msk=missing)

Define initial condition (u0,du) using randomly distributed longitude, latitude positions on the sphere (randn_lonlat).

IndividualDisplacements.postprocess_ODESolution — Method

postprocess_ODESolution()

Copy sol to a DataFrame & map position to lon,lat coordinates

IndividualDisplacements.postprocess_ODESolution_simple — Method

postprocess_ODESolution_simple()

Copy sol to a DataFrame & map position to x,y coordinates, and define time axis for a simple doubly periodic domain

IndividualDisplacements.randn_lonlat — Function

randn_lonlat(nn=1,seed=missing)

Randomly distributed longitude, latitude positions on the sphere.

IndividualDisplacements.read_drifters — Method

read_drifters(pth,lst;chnk=Inf,rng=(missing,missing))

Read near-surface drifter data (https://doi.org/10.1002/2016JC011716) from the Global Drifter Program (https://doi.org/10.25921/7ntx-z961) into a DataFrame

Note: need to use NetCDF.jl as NCDatasets.jl errors when TIME = Inf

pth="Drifter_hourly_v013/"
lst=["driftertrajGPS_1.03.nc","driftertrajWMLE_1.02_block1.nc","driftertrajWMLE_1.02_block2.nc",
   "driftertrajWMLE_1.02_block3.nc","driftertrajWMLE_1.02_block4.nc","driftertrajWMLE_1.02_block5.nc",
   "driftertrajWMLE_1.02_block6.nc","driftertrajWMLE_1.03_block7.nc"]

#df=read_drifters( pth*lst[end],chnk=1000,rng=(2014.1,2014.2) )

df = DataFrame(lon=[], lat=[], t=[], ID=[])
for fil in lst
   println(fil)
   append!(df,read_drifters( pth*fil,chnk=10000,rng=(2005.0,2020.0) ))
   println(size(df))
end

#sort!(df, [:t, :lat])
#CSV.write(pth*"Drifter_hourly_2005_2019.csv", df)
#unID=unique(df[!,:ID])

IndividualDisplacements.read_flt — Method

read_flt(dirIn::String,prec::DataType)

Read displacements from MITgcm/pkg/flt output file into a DataFrame.

IndividualDisplacements.read_mds — Method

read_mds(filRoot::String,x::MeshArray)

Read a gridded variable from 2x2 tile files. This is used in example2_setup() with flt_example/

IndividualDisplacements.read_uvetc — Method

read_uvetc(k::Int,γ::Dict,pth::String)

Define uvetc given the grid variables γ and a vertical level choice k including velocities obtained from files in pth

IndividualDisplacements.read_uvetc — Method

read_uvetc(k::Int,t::Float64,γ::Dict,pth::String)

Define uvetc given the grid variables γ, a vertical level choice k, the time t in seconds (Float64), and velocities obtained from files in pth.

The two climatological months (m0,m1) that bracket time t will be extracted (e.g. months 12 & 1 then 1 & 2 and so on).

Note: the nitial implementation does this only approximately by setting every months duration to 1 year / 12 for simplicity; should be improved...

IndividualDisplacements.NeighborTileIndices_cs — Method

NeighborTileIndices_cs(grid::Dict)

Derive list of neighboring tile indices for a cs or llc grid + functions that convert indices from one tile to another. Returns a Dict to merge later.

IndividualDisplacements.NeighborTileIndices_dpdo — Method

NeighborTileIndices_dpdo(ni::Int,nj::Int)

List of W, E, S, N neighbor tile IDs in the case of a doubly periodic domain with ni x nj tiles.

IndividualDisplacements.RelocationFunctions_cs — Method

RelocationFunctions_cs(xmpl)

Define matrix of functions to convert indices across neighboring tiles

IndividualDisplacements.RelocationFunctions_cs_check — Method

RelocationFunctions_cs_check(xmpl,RF,trgt)

Visualize that RelocationFunctions_cs behaves as expected

IndividualDisplacements.VelComp! — Method

VelComp!(du,u,p::Dict,tim)

Interpolate velocity from gridded fields (after exchange on u0,v0) and return position increment du (i.e. x,y,fIndex).

IndividualDisplacements.VelComp — Method

VelComp(du,u,p::Dict,tim)

Interpolate velocity from gridded fields and return position increment du

IndividualDisplacements.VelCopy — Method

VelCopy(du,u,p::DataFrame,t)

Interpolate velocity from MITgcm float_trajectories output and return position increment du.

IndividualDisplacements.read_velocities — Method

read_velocities(γ::gcmgrid,t::Int,pth::String)

Read velocity components u,v from files in pthfor time t

IndividualDisplacements.update_location_cs! — Method

update_location_cs!

Update location (x,y,fIndex) when out of domain. Note: initially, this only works for the dpdo grid type provided by MeshArrays.jl.

IndividualDisplacements.update_location_dpdo! — Method

update_location_dpdo!

Update location (x,y,fIndex) when out of domain. Note: initially, this only works for the dpdo grid type provided by MeshArrays.jl.