IndividualDisplacements.jl

IndividualDisplacements.jl computes elementary point displacements over a gridded model domain. It can also read / write them from / to file. A typical application is the simulation and analysis of materials drifting or moving over the Global Ocean (e.g. plastics or planktons) or Atmosphere (e.g. dust or chemicals). Inter-operability with popular climate model grids and MeshArrays.jl is an important prospect. IndividualDisplacements.jl was initially designed in relation to MITgcm, ECCOv4 (Forget et al. 2015), and CBIOMES model simulations.

API Guide

IndividualDisplacements.NeighborTileIndices_dpdo

IndividualDisplacements.ReadDisplacements

IndividualDisplacements.ReadGriddedFields

IndividualDisplacements.VelComp

IndividualDisplacements.VelComp!

IndividualDisplacements.VelCopy

IndividualDisplacements.ex_1

IndividualDisplacements.ex_2

IndividualDisplacements.myread

IndividualDisplacements.ReadDisplacements — Method

ReadDisplacements(dirIn::String,prec::DataType)

Read displacements from MITgcm output file using MeshArrays and return as a DataFrame.

IndividualDisplacements.ReadGriddedFields — Method

ReadGriddedFields()

Read gridded variables from file using MeshArrays and return result in uvetc Dictionary.

IndividualDisplacements.VelComp! — Method

VelComp!(du,u,p::Dict,tim)

Interpolate velocity from gridded fields and return position increment du

IndividualDisplacements.VelComp — Method

VelComp(du,u,p::Dict,tim)

Interpolate velocity from gridded fields and return position increment du

IndividualDisplacements.VelCopy — Method

VelCopy(du,u,p::DataFrame,t)

Interpolate velocity from MITgcm float_trajectories output and return position increment du.

IndividualDisplacements.NeighborTileIndices_dpdo — Method

NeighborTileIndices_dpdo(ni::Int,nj::Int)

List of W, E, S, N neighbor tile IDs in the case of a doubly periodic domain with ni x nj tiles.

IndividualDisplacements.ex_1 — Method

ex_1()

Global ocean case – just reading from file for now.

df=IndividualDisplacements.ex_1()

p=dirname(pathof(IndividualDisplacements))
include(joinpath(p,"PlotIndDisp.jl"))
PyPlot.figure(); PlotMapProj(df,300); gcf()

IndividualDisplacements.ex_2 — Method

ex_2()

Reproducing MITgcm/verification/flt_example/ case. This is based on an extended and modified configuration of the standard MITgcm test case.

(df,ref,sol)=IndividualDisplacements.ex_2();

p=dirname(pathof(IndividualDisplacements))
include(joinpath(p,"PlotIndDisp.jl"))
PyPlot.figure(); PlotBasic(df,300,100000.0); gcf()

using Plots
Plots.plot(sol[1,:],sol[2,:],linewidth=5,lc=:black, title="One Trajectory Example",
xaxis="x",yaxis="y",label="Julia Solution") # legend=false
pl=Plots.plot!(ref[1,:],ref[2,:],lw=3,ls=:dash,lc=:red,label="MITgcm Solution")

IndividualDisplacements.myread — Method

myread()

Read a gridded variable from 2x2 tile files. This is used in ReadGriddedFields() with flt_example/