API Guide

The ⬡ and ⬡! functions compute the tracked point / individual / agent velocities.

IndividualDisplacements.⬡! — Function

VelComp!(du,u,p::Dict,tim)

Interpolate velocity from gridded fields (after exchange on u0,v0) and return position increment du (i.e. x,y,fIndex).

IndividualDisplacements.⬡ — Function

VelComp(du,u,p::Dict,tim)

Interpolate velocity from gridded fields and return position increment du

Setup And Postprocessing

Convenience functions to initialize a simulation and posprocess the output are provided.

IndividualDisplacements.setup_periodic_domain — Function

setup_periodic_domain(np::Integer=16)

Set up a periodic domain of size np x np

np=16 #domain size is np x np
Γ=setup_periodic_domain(np)

IndividualDisplacements.initialize_randn — Function

initialize_randn(Γ::Dict,n::Int=1 ; s=missing,msk=missing)

Define initial condition (u0,du) using randomly distributed longitude, latitude positions on the sphere (randn_lonlat).

IndividualDisplacements.initialize_gridded — Function

initialize_gridded(uvetc::Dict,n_subset::Int=1)

Define initial condition (u0,du) as a subset of grid points

IndividualDisplacements.postprocess_lonlat — Function

postprocess_lonlat()

Copy sol to a DataFrame & map position to lon,lat coordinates

IndividualDisplacements.postprocess_xy — Function

postprocess_xy()

Copy sol to a DataFrame & map position to x,y coordinates, and define time axis for a simple doubly periodic domain

Read Output From File

Trajectory simulated by the MITgcm or observed by the global drifter program can be read from file.

IndividualDisplacements.read_flt — Function

read_flt(dirIn::String,prec::DataType)

Read displacements from MITgcm/pkg/flt output file into a DataFrame.

IndividualDisplacements.read_drifters — Function

read_drifters(pth,lst;chnk=Inf,rng=(missing,missing))

Read near-surface drifter data (https://doi.org/10.1002/2016JC011716) from the Global Drifter Program (https://doi.org/10.25921/7ntx-z961) into a DataFrame

Note: need to use NetCDF.jl as NCDatasets.jl errors when TIME = Inf

pth="Drifter_hourly_v013/"
lst=["driftertrajGPS_1.03.nc","driftertrajWMLE_1.02_block1.nc","driftertrajWMLE_1.02_block2.nc",
   "driftertrajWMLE_1.02_block3.nc","driftertrajWMLE_1.02_block4.nc","driftertrajWMLE_1.02_block5.nc",
   "driftertrajWMLE_1.02_block6.nc","driftertrajWMLE_1.03_block7.nc"]

#df=read_drifters( pth*lst[end],chnk=1000,rng=(2014.1,2014.2) )

df = DataFrame(lon=[], lat=[], t=[], ID=[])
for fil in lst
   println(fil)
   append!(df,read_drifters( pth*fil,chnk=10000,rng=(2005.0,2020.0) ))
   println(size(df))
end

#sort!(df, [:t, :lat])
#CSV.write(pth*"Drifter_hourly_2005_2019.csv", df)
#unID=unique(df[!,:ID])

Types